Windows cleanup.
