+++ src/bear_main.h
@@ -4,8 +4,8 @@
 #include "glad.h"
 
 #include "bear_types.h"
-#include "bear_audio.h"
-#include "bear_ecs.h"
+#include "audio/bear_audio.h"
+#include "ecs/bear_ecs.h"
 
 struct OSFile
 {
@@ -28,6 +28,12 @@ struct PLT
 	int32 (*last_write) (const char *);
 };
 
+struct CLK
+{
+	float64 time;
+	float32 delta;
+};
+
 struct MemoryAllocation
 {
 	const char *file;
@@ -43,6 +49,9 @@ struct World
 		bool jump;
 	} input;
 
+	// A clock for timing.
+	CLK clk;
+
 	// Platform functions.
 	PLT plt;
 
@@ -51,6 +60,8 @@ struct World
 
 	// Audio, so the world can feel the beats.
 	Audio audio;
+
+	bool running;
 
 	// TODO: Remove in reloase
 	uint32 __mem_length = 0;
@@ -71,6 +82,6 @@ struct World
 
 #endif
 
-#include "bear_ecs_init.cpp"
+#include "ecs/bear_ecs_init.cpp"
