+++ src/bear_types.h
@@ -1,10 +1,10 @@
 // Primitive types
 typedef const char *string;
 
-typedef char	int8;
-typedef short	int16;
-typedef int		int32;
-typedef long	int64;
+typedef char		int8;
+typedef short		int16;
+typedef int			int32;
+typedef long		int64;
 
 typedef unsigned char	uint8;
 typedef unsigned short	uint16;
