+++ src/bear_windows.cpp
@@ -13,15 +13,23 @@ int win_printf(const char *format, ...);
 typedef void (*StepFunc)(World *, float32);
 typedef void (*SoundFunc)(int16 *, int32);
 
-StepFunc game_step;
-SoundFunc game_sound;
-bool lock_sound = false;
+struct GameHandle
+{
+	StepFunc step;
+	SoundFunc sound;
+
+	// So the sound thread doesn't run when we unload.
+	SDL_mutex *lock;
+
+	int32 access_time;
+	HMODULE lib;
+};
+
+static GameHandle game;
 
 // This is ugly... I know.
 MemoryAllocation __mem[1024];
 
-int32 last_access_time = 0;
-HMODULE handle;
 
 int32 win_printf(const char *format, ...)
 {
@@ -62,52 +70,54 @@ WINDOWS_FILE_ERROR:
 	return -1;
 }
 
-bool load_libbear()
+bool load_libbear(GameHandle *handle)
 {
+	GameHandle game = *handle;
 	const char *path = "bin/libbear.dll";
 	const char *temp_path = "bin/libbear_temp.dll";
-	int32 last = get_file_edit_time(path);
-	if (last == last_access_time)
+	int32 access_time = get_file_edit_time(path);
+	if (access_time == game.access_time)
 	{
 		return false;
 	}
-	if (handle)
+
+	SDL_LockMutex(game.lock);
+	if (game.lib)
 	{
-		FreeLibrary(handle);
+		FreeLibrary(game.lib);
 	}
 	CopyFile(path, temp_path, false);
-	handle = LoadLibrary(temp_path);
-	if (!handle)
+	game.lib = LoadLibrary(temp_path);
+	if (!game.lib)
 	{
-		win_printf("[DL-ERROR] %d\n", GetLastError());
+		win_printf("[DLL-ERROR] %d\n", GetLastError());
 		DEBUG_LOG("Failed to load libgame.so!");
 		return false;
 	}
-	StepFunc step = (StepFunc) GetProcAddress(handle, "step");
+	StepFunc step = (StepFunc) GetProcAddress(game.lib, "step");
 	if (!step)
 	{
-		win_printf("[DL-ERROR] %d\n", GetLastError());
+		win_printf("[DLL-ERROR] %d\n", GetLastError());
 		DEBUG_LOG("Failed to load game_update!");
 		return false;
 	}
-	SoundFunc sound = (SoundFunc) GetProcAddress(handle, "sound");
+	SoundFunc sound = (SoundFunc) GetProcAddress(game.lib, "sound");
 	if (!sound)
 	{
-		win_printf("[DL-ERROR] %d\n", GetLastError());
+		win_printf("[DLL-ERROR] %d\n", GetLastError());
 		DEBUG_LOG("Failed to load game_draw!");
 		return false;
 	}
 
 	DEBUG_LOG("Reload!");
-	last_access_time = last;
+	game.access_time = access_time;
 
-	game_step = step;
+	game.step = step;
 
-	while (lock_sound);
-	lock_sound = true;
-	game_sound = sound;
-	lock_sound = false;
+	game.sound = sound;
+	SDL_UnlockMutex(game.lock);
 
+	*handle = game;
 	return true;
 }
 
@@ -148,22 +158,16 @@ void free_file(OSFile file)
 
 void plt_audio_callback(void *userdata, uint8 *stream, int32 length)
 {
-	while (lock_sound);
-	lock_sound = true;
-	game_sound((int16 *) stream, length / 2);
-
-	lock_sound = false;
+	SDL_LockMutex(game.lock);
+	game.sound((int16 *) stream, length / 2);
+	SDL_UnlockMutex(game.lock);
 }
 
-#ifdef asdas //__DEBUG 
-int main(int varc, char *varv[])
-#else
 int CALLBACK WinMain(
   HINSTANCE hInstance,
   HINSTANCE hPrevInstance,
   LPSTR     lpCmdLine,
   int       nCmdShow)
-#endif
 {
 	world.plt.malloc = malloc_;
 	world.plt.free = free_;
@@ -183,13 +187,14 @@ int CALLBACK WinMain(
 
 	init_ecs(&world);
 
-	if (load_libbear() == false)
+	game.lock = SDL_CreateMutex();
+	if (load_libbear(&game) == false)
 	{
 		return(-1);
 	}
 	//has_file_changed();
 
-	if (SDL_Init(SDL_INIT_EVERYTHING) != 0)
+	if (SDL_Init(SDL_INIT_AUDIO | SDL_INIT_VIDEO | SDL_INIT_GAMECONTROLLER) != 0)
 	{
 		DEBUG_LOG("Unable to initalize SDL.");
 		SDL_Quit();
@@ -213,10 +218,11 @@ int CALLBACK WinMain(
 	// TODO: Use OpenGL 3.3, or newer. This is just to get hello triangle.
 	SDL_GL_SetAttribute(SDL_GL_CONTEXT_MAJOR_VERSION, 2);
 	SDL_GL_SetAttribute(SDL_GL_CONTEXT_MINOR_VERSION, 1);
-	SDL_GL_SetSwapInterval(1);
 	SDL_GL_SetAttribute(SDL_GL_DOUBLEBUFFER, 1);
 	SDL_GL_SetAttribute(SDL_GL_DEPTH_SIZE, 24);
 	SDL_GL_CreateContext(window);
+
+	SDL_GL_SetSwapInterval(0);
 
 	if (gladLoadGL() == 0)
 	{
@@ -259,17 +265,28 @@ int CALLBACK WinMain(
 	
 	DEBUG_LOG("Window launch!");
 
-	bool running = true;
-	while (running)
+	world.running = true;
+
+	LARGE_INTEGER counter_frequency = counter_frequency;
+	QueryPerformanceFrequency(&counter_frequency);
+	LARGE_INTEGER counter;
+	LARGE_INTEGER last_counter;
+	LARGE_INTEGER start;
+	QueryPerformanceCounter(&start);
+ 
+	QueryPerformanceCounter(&last_counter);
+	while (world.running)
 	{
-		load_libbear();
+		
+
+		load_libbear(&game);
 
 		SDL_Event event;
 		while (SDL_PollEvent(&event))
 		{
 			if (event.type == SDL_QUIT)
 			{
-				running = false;
+				world.running = false;
 			}
 			if (event.type == SDL_KEYDOWN)
 			{
@@ -281,10 +298,18 @@ int CALLBACK WinMain(
 			}
 		}
 		
-		game_step(&world, 0.1f);
+		game.step(&world, world.clk.delta);
 
 		SDL_GL_SwapWindow(window);
+		
+		// Timing
+		QueryPerformanceCounter(&counter);
+		int64 delta_counter = counter.QuadPart - last_counter.QuadPart;
+		last_counter = counter;
+		world.clk.delta = (float64) delta_counter / (float64) counter_frequency.QuadPart;
+		world.clk.time = (float64) (counter.QuadPart - start.QuadPart) / (float64) counter_frequency.QuadPart;
 	}
+	SDL_DestroyMutex(game.lock);
 	SDL_CloseAudio();
 	SDL_Quit();
