--- src/ecs/bear_component.h
@@ -1,5 +1,0 @@
-#pragma once
-
-
-add_system(system);
-add_component(entity, component); // -> Find the system(s), add it to them.
