+++ src/ecs/bear_ecs_init.cpp
