+++ src/math/bear_math.h
@@ -9,6 +9,7 @@
 //
 
 #include <math.h>
+#define PI 3.1419f
 
 // Custom Math functions.
 template <typename T>
